backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
redis[hiredis]>=5.0.0
pyyaml>=6.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
apscheduler>=3.10.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
